Refactoring, Cleanup, things that need to change in migrations generator
